Implementation of TokenAcquirerFactory in ASPNETCore + tests
