+++ src/Microsoft.Identity.Web.TokenAcquisition/ServiceCollectionExtensions.cs
@@ -64,7 +64,8 @@ namespace Microsoft.Identity.Web
             ServiceDescriptor? tokenAcquisitionInternalService = services.FirstOrDefault(s => s.ServiceType == typeof(ITokenAcquisitionInternal));
             ServiceDescriptor? tokenAcquisitionhost = services.FirstOrDefault(s => s.ServiceType == typeof(ITokenAcquisitionHost));
             ServiceDescriptor? authenticationHeaderCreator = services.FirstOrDefault(s => s.ServiceType == typeof(IAuthorizationHeaderProvider));
-            if (tokenAcquisitionService != null && tokenAcquisitionInternalService != null && tokenAcquisitionhost != null && authenticationHeaderCreator != null)
+            ServiceDescriptor? tokenAcquirerFactory = services.FirstOrDefault(s => s.ServiceType == typeof(ITokenAcquirerFactory));
+            if (tokenAcquisitionService != null && tokenAcquisitionInternalService != null && tokenAcquisitionhost != null && authenticationHeaderCreator != null )
             {
                 if (isTokenAcquisitionSingleton ^ (tokenAcquisitionService.Lifetime == ServiceLifetime.Singleton))
                 {
@@ -73,6 +74,7 @@ namespace Microsoft.Identity.Web
                     services.Remove(tokenAcquisitionInternalService);
                     services.Remove(tokenAcquisitionhost);
                     services.Remove(authenticationHeaderCreator);
+                    services.Remove(tokenAcquirerFactory);
                 }
                 else
                 {
@@ -88,8 +90,10 @@ namespace Microsoft.Identity.Web
                 // ASP.NET Core
                 services.AddHttpContextAccessor();
                 services.AddSingleton<ITokenAcquisition, TokenAcquisitionAspNetCore>();
+                services.AddSingleton(s => (ITokenAcquirerFactory)s.GetRequiredService<ITokenAcquisition>());
 
                 services.AddSingleton<ITokenAcquisitionHost, TokenAcquisitionAspnetCoreHost>();
+
 #else
                 // .NET FW.
                 services.AddSingleton<ITokenAcquisition, TokenAcquisition>();
@@ -105,6 +109,7 @@ namespace Microsoft.Identity.Web
                 services.AddHttpContextAccessor();
 
                 services.AddScoped<ITokenAcquisition, TokenAcquisitionAspNetCore>();
+                services.AddScoped(s => (ITokenAcquirerFactory)s.GetRequiredService<ITokenAcquisition>());
 
                 services.AddScoped<ITokenAcquisitionHost, TokenAcquisitionAspnetCoreHost>();
 #else
