+++ tests/WebAppCallsMicrosoftGraph/appsettings.json
@@ -13,10 +13,10 @@
         // with a priority order, which allows for adaptive configuration
         // depending on the deployment
         "ClientCredentials": [
-            {
-                "SourceType": "SignedAssertionFromManagedIdentity",
-                "ManagedIdentityClientId": ""
-            },
+            //{
+            //    "SourceType": "SignedAssertionFromManagedIdentity",
+            //    "ManagedIdentityClientId": ""
+            //},
             {
                 "SourceType": "KeyVault",
                 "KeyVaultUrl": "https://webappsapistests.vault.azure.net",
